requirements.txt
discord.py==2.5.2
python-dotenv==1.1.1
aiohttp==3.12.13
